loading trained umap if present

+++ src/dimensionality_reduction.py
@@ -26,7 +26,7 @@ import numexpr
 import time
 
 from src.utils.function_runner import run_function_with_overrides, execute_with_gpu_logging
-from src.utils.utils import load_h5py, load_with_indices_h5py_efficient, get_indices_for_random_h5py_subset, save_h5py, load_with_indices_h5py
+from src.utils.utils import load_h5py, load_with_indices_h5py_efficient, get_indices_for_random_h5py_subset, save_h5py, load_with_indices_h5py, get_number_of_samples_h5py
 
 def UMAP_transform_full_fit(
     PROCESSED_REDDIT_DATA: str,
@@ -62,36 +62,42 @@ def UMAP_transform_partial_fit(
     
     """
     Load embeddings, sample a subset, fit UMAP on the subset, save the model, and transform the entire dataset.
+    If the model is already present in the designated location, load it instead of training a new one.
     """
 
-    print("subset size for partial fit", PARTIAL_FIT_DIM_REDUCTION)
-    
-    umap_model = UMAP(
-        n_neighbors=UMAP_N_Neighbors,
-        n_components=UMAP_COMPONENTS,
-        min_dist=UMAP_MINDIST,
-        negative_sample_rate=NEGATIVE_SAMPLE_RATE,
-        n_epochs=UMAP_N_EPOCHS,
-    )
+    if os.path.exists(UMAP_MODEL_SAVE_PATH):
+        logger.info(f"Loading existing UMAP model from {UMAP_MODEL_SAVE_PATH}")
+        umap_model = joblib.load(UMAP_MODEL_SAVE_PATH)
+    else:
+        print("subset size for partial fit", PARTIAL_FIT_DIM_REDUCTION)
+        
+        umap_model = UMAP(
+            n_neighbors=UMAP_N_Neighbors,
+            n_components=UMAP_COMPONENTS,
+            min_dist=UMAP_MINDIST,
+            negative_sample_rate=NEGATIVE_SAMPLE_RATE,
+            n_epochs=UMAP_N_EPOCHS,
+        )
 
-    partial_fit_indices, total_samples, num_samples = get_indices_for_random_h5py_subset(PROCESSED_REDDIT_DATA, "embeddings", PARTIAL_FIT_DIM_REDUCTION)
-    logger.info(f"Running partial fit on {num_samples} samples out of {total_samples} samples")
-    s = time.time()
+        partial_fit_indices = get_indices_for_random_h5py_subset(PROCESSED_REDDIT_DATA, "embeddings", PARTIAL_FIT_DIM_REDUCTION)
+        logger.info(f"Running partial fit on {num_samples} samples out of {total_samples} samples")
+        s = time.time()
 
-    # here we only load the necessary indices otherwise oom error
-    sampled_features = load_with_indices_h5py(PROCESSED_REDDIT_DATA, "embeddings", partial_fit_indices)  
-    logger.info(f"Time to load data: {time.time() - s:.2f} s")
-    execute_with_gpu_logging(umap_model.fit, sampled_features)
+        # here we only load the necessary indices otherwise oom error
+        sampled_features = load_with_indices_h5py(PROCESSED_REDDIT_DATA, "embeddings", partial_fit_indices)  
+        logger.info(f"Time to load data: {time.time() - s:.2f} s")
+        execute_with_gpu_logging(umap_model.fit, sampled_features)
 
-    # Save the fitted UMAP model
-    joblib.dump(umap_model, UMAP_MODEL_SAVE_PATH)
-    logger.info(f"UMAP model saved at {UMAP_MODEL_SAVE_PATH}")
+        # Save the fitted UMAP model
+        joblib.dump(umap_model, UMAP_MODEL_SAVE_PATH)
+        logger.info(f"UMAP model saved at {UMAP_MODEL_SAVE_PATH}")
 
     # iterate over the rest of the data in chunks of subset_size and transform
     # subset size (derived from PARTIAL_FIT_DIM_REDUCTION) should be the maximum subset of data on which we can fit
     # given a certain GPU memory 
+    total_samples, num_samples = get_number_of_samples_h5py(PROCESSED_REDDIT_DATA, "embeddings", PARTIAL_FIT_DIM_REDUCTION)
     result = None
-    for i in tqdm(range(0, total_samples, num_samples//2)): # transform takes more memory than fit
+    for i in tqdm(range(0, total_samples, num_samples//4)): # transform takes more memory than fit
         indices = np.arange(i, min(i + num_samples, total_samples))
         chunk = load_with_indices_h5py_efficient(PROCESSED_REDDIT_DATA, "embeddings", indices)
         transformed_chunk = execute_with_gpu_logging(umap_model.transform, chunk)
@@ -101,7 +107,6 @@ def UMAP_transform_partial_fit(
             result = np.concatenate((result, transformed_chunk), axis=0)
 
     save_h5py(result, PROCESSED_REDDIT_DATA, DIMENSIONALITY_REDUCTION_DB_NAME)
-
 
 if __name__ == "__main__":
     print("Total running time:", run_function_with_overrides(UMAP_transform_partial_fit, config))
